remove labels from series runtime_function_executions_num (#163)
In observability, it is a really bad idea to create a metric
that has labels with unbounded number of values. Initially, I had
added plugin_name and function_name as labels to the metric
runtime_function_executions_num. Both of them can increase to
a large number for each pod as time progresses. This PR removes
both of these labels for now and we can add them back if needed.