correctly reset the timer when it is not fired for inference history
We periodically send inference history data to postgres.
The periodicity depends on both time and amount of data.
We use a timer to ensure that data is being sent to postgres
for the case when not a lot of data is being generated.

But if data is continuously being generated, we were
resetting the timer without draining the timer channel.
The effect of this would be that timers will get fired
later on without waiting the right amount of time.

This commit fixes that. This is being fixed in Go 1.23
and we can simply call timer.Reset() in the begining of
the loop irrespective of the state of the timer.